Support JSX namespaced name pass-through
The Draft JSX Specification includes syntax for namespaced names for
elements and attributes in the form of `<nsName:elementName />` and
`<Component nsName:attributeName />`.
https://facebook.github.io/jsx/#syntax

This change enables the use of JSX namespaced names in .tsx when JSX
code generation is configured to `preserve`. The .jsx output may then be
post-processed using other tools.